found the maximum from three float and passed three given test cases

+++ src/main/java/com/bz/findmax/FindMax.java
@@ -2,8 +2,8 @@ package com.bz.findmax;
 import java.lang.*;
 public class FindMax {
 
-    public int findMaximun(Integer num1 , Integer num2 , Integer num3){
-        int max = num1;
+    public Float findMaximun(Float num1 , Float num2 , Float num3){
+        Float max = num1;
         if(num2.compareTo(max)>0){
             max=num2;
         }
